streaming/pyproject.toml
[project]
name = "streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.0",
    "python-dotenv>=1.1.1",
]
